[CHG] Split skii.platform.factories in 4 files
    Factories was all in the same file. Dispatch them on
    agent/event/resource/common module and give short/unique import path
    with __all__ = [...] in package __init__ module.
    Also start flake8/pyright inspection/fix
